Handle missing resources on refresh
Gracefully handle the situation where a resource has gone missing
outside of Terraform (e.g., because the user manually ran a `DROP`
command).

This logic was applied to table resources as a one-off in #155, but that
code appears to have gotten lost in 9235b4b. This commit re-adds the
logic for all resource types.

No tests, because our current testing infrastructure is not powerful
enough to handle this. Will put up a test in a separate PR that enhances
the testing infrastructure (see #165).

Fix #157.