Fix identifier quoting
This commit fixes the follow-up to #233, in which role grants to roles
that contained dots (like `joe@materialize.com`) were reported to fail.

The cause was that the `QuoteIdentifier` function was incorrectly
splitting identifiers on `.` and then escaping each component. As far as
I can tell, this code is cruft left over from an earlier version of the
Terraform provider, which was less rigorous in its handling of IDs. Just
removing the splitting fixes the bug and causes no other regressions in
the test suite.